+++ app/controllers/profiles_controller.rb
@@ -3,6 +3,7 @@
 # Controller for profiles, such as users
 class ProfilesController < ApplicationController
   def show
-    @profile = Handle.find_by_identifier!(params[:handle]).profile
+    @profile = Profile.find params[:handle]
+    @projects = @profile.projects.order id: :desc
   end
 end