+++ app/helpers/layout_helper.rb
@@ -20,9 +20,10 @@ module LayoutHelper
   def navigation_links
     if account_signed_in?
       links = [
-        { text: 'Profile',  path: profile_path(current_account.user) },
-        { text: 'Account',  path: edit_account_path },
-        { text: 'Logout',   path: destroy_session_path }
+        { text: 'New Project',  path: new_project_path },
+        { text: 'Profile',      path: url_for(current_account.user) },
+        { text: 'Account',      path: edit_account_path },
+        { text: 'Logout',       path: destroy_session_path }
       ]
     else
       links = [