+++ spec/models/user_spec.rb
@@ -7,10 +7,22 @@ RSpec.describe User, type: :model do
     is_expected.to be_valid
   end
 
+  describe 'route keys' do
+    it 'should have singural route key: profile' do
+      expect(user.model_name.singular_route_key).to eq 'profile'
+    end
+    it 'should have route key: profiles' do
+      expect(user.model_name.route_key).to eq 'profiles'
+    end
+  end
+
   describe 'associations' do
     it { is_expected.to belong_to(:account) }
     it do
       is_expected.to have_one(:handle).dependent(:destroy).inverse_of :profile
+    end
+    it do
+      is_expected.to have_many(:projects).dependent(:destroy).inverse_of :owner
     end
   end