+++ spec/rails_helper.rb
@@ -40,6 +40,7 @@ require 'database_cleaner'
 #
 # Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
 require Rails.root.join('spec', 'support', 'database_cleaner.rb')
+require Rails.root.join('spec', 'support', 'tmp_file_cleaner.rb')
 require Rails.root.join('spec', 'support', 'helpers', 'features_helper.rb')
 
 # Checks for pending migration and applies them before tests are run.
@@ -71,6 +72,13 @@ RSpec.configure do |config|
     end
   end
 
+  # Enable partial rendering from application folder
+  # Fixes https://github.com/rspec/rspec-rails/issues/396
+  # Solution: https://github.com/verypossible/raygun-rails-template/pull/14
+  config.before(:example, type: :view) do
+    view.lookup_context.view_paths.unshift 'app/views/application'
+  end
+
   # RSpec Rails can automatically mix in different behaviours to your tests
   # based on their file location, for example enabling you to call `get` and
   # `post` in specs under `spec/controllers`.
@@ -90,6 +98,9 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
+
+  # Include Devise test helpers
+  config.include Devise::Test::ControllerHelpers, type: :controller
 
   # Include Feature test helpers
   config.include FeaturesHelper, type: :feature